ril: Support new sound manger RIL responses
New version of Samsung RIL do not use RIL_UNSOL_WB_AMR_STATE anymore.
There is a Sound Manager (SND-MGR) which has its own response base
(20000). So to be able to register a handler for

RIL_UNSOL_SNDMGR_WB_AMR_REPORT

with libsecril-client we need to know about it. This allows us to
register a callback in the audio HAL to turn WB support on.

WB_AMR_REPORT report 0x00, 0x01 or 0x02. I guess 0x01 is WB ON and
0x02 is WB + NS ON.

Change-Id: I403c86781551fc3b08354187322e8443402d56f6